deprecation (#247)
* prepare for 64 bit

only linker errors for libcurl and cocos2d are to fix (need to get the new libs)

* click between frames warning

* improve best run

* add more labels

* Update gd.hpp

* Update CHANGELOG.md

* ui tweaks

* fix 64 bit crashhandler

* x64 libcurl

* Update gd.hpp

* remove 4gb patch

* 2.206 changes

* update libs

* x64 workflow

* update libs

* 2.206

* "small fix"

* tiny clean-ups

* random edits

* Update README.md

+++ src/geode/hooks/PlayLayer.cpp
@@ -4,6 +4,7 @@
 #include "../../shared/hacks/auto-pickup-coins/auto-pickup-coins.hpp"
 #include "../../shared/hacks/startpos-switcher/startpos-switcher.hpp"
 #include "../../shared/hacks/labels/labels.hpp"
+#include "../../shared/hacks/noclip/noclip.hpp"
 #include "../../shared/hacks/noclip-limit/noclip-limit.hpp"
 #include "../../shared/hacks/zephyrus/replays.hpp"
 #include "../../shared/hacks/random-seed/random-seed.hpp"
@@ -17,8 +18,12 @@
 
 namespace openhack::hooks {
     struct PlayLayerHook : geode::Modify<PlayLayerHook, PlayLayer> {
+        static void onModify(auto& self) {
+            (void) self.setHookPriority("PlayLayer::destroyPlayer", 0x500000);
+        }
+
         bool init(GJGameLevel *level, bool useReplay, bool dontCreateObjects) {
-            hacks::Display::playLayerInit(reinterpret_cast<gd::GJGameLevel *>(level));
+            hacks::Display::playLayerInit(level);
             hacks::AutoPickupCoins::initLevel();
             hacks::StartPosSwitcher::initLevel();
             hacks::Labels::playLayerInit();
@@ -49,19 +54,20 @@ namespace openhack::hooks {
 
         void addObject(GameObject *object) {
             PlayLayer::addObject(object);
-            hacks::AutoPickupCoins::addObject(reinterpret_cast<gd::GameObject *>(object));
-            hacks::StartPosSwitcher::addObject(reinterpret_cast<gd::GameObject *>(object));
-            hacks::SmartStartPos::addObject(reinterpret_cast<gd::GameObject *>(object));
+            hacks::AutoPickupCoins::addObject(object);
+            hacks::StartPosSwitcher::addObject(object);
+            hacks::SmartStartPos::addObject(object);
         }
 
-        void destroyPlayer(PlayerObject* player, GameObject* object) {
-            hacks::NoclipLimit::destroyPlayer(reinterpret_cast<gd::GameObject *>(object));
-            PlayLayer::destroyPlayer(player, object);
+        void destroyPlayer(PlayerObject* player, GameObject* object) override {
+            hacks::NoclipLimit::destroyPlayer(player, object);
+            if (!hacks::Noclip::destroyPlayer(player, object))
+                PlayLayer::destroyPlayer(player, object);
             hacks::NoclipLimit::postDestroyPlayer();
             hacks::Hitboxes::destroyPlayer();
         }
 
-        void postUpdate(float dt) {
+        void postUpdate(float dt) override {
             PlayLayer::postUpdate(dt);
             hacks::Hitboxes::postUpdate();
             hacks::AccuratePercentage::postUpdate();
@@ -71,6 +77,7 @@ namespace openhack::hooks {
         void fullReset() {
             PlayLayer::fullReset();
             hacks::Hitboxes::fullReset();
+            config::setGlobal<int>("bestRun", 0);
         }
 
         void playEndAnimationToPos(cocos2d::CCPoint pos) {
